Replace the "CommonConfig" with "MainConfig"
