+++ circuit/src/gadgets/mod.rs
@@ -4,7 +4,5 @@ pub mod bits2num;
 pub mod common;
 /// Check if a number is less than or equal to a value
 pub mod lt_eq;
-/// Main gate to be used for other gates
-pub mod main;
 /// Set membership gadget
 pub mod set;
