+++ circuit/src/lib.rs
@@ -157,9 +157,9 @@ impl<'a, F: FieldExt> RegionCtx<'a, F> {
 }
 
 /// Number of advice columns in common config
-pub const ADVICE: usize = 7;
+pub const ADVICE: usize = 5;
 /// Number of fixed columns in common config
-pub const FIXED: usize = 10;
+pub const FIXED: usize = 8;
 
 /// Common config for the whole circuit
 #[derive(Clone, Debug)]
